# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * To-Do List Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  userId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus =
  (typeof TaskInputStatus)[keyof typeof TaskInputStatus];

export const TaskInputStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type TaskInputPriority =
  (typeof TaskInputPriority)[keyof typeof TaskInputPriority];

export const TaskInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  priority: TaskInputPriority;
  dueDate?: string;
}

export type TaskUpdateStatus =
  (typeof TaskUpdateStatus)[keyof typeof TaskUpdateStatus];

export const TaskUpdateStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type TaskUpdatePriority =
  (typeof TaskUpdatePriority)[keyof typeof TaskUpdatePriority];

export const TaskUpdatePriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardStats {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  highPriority: number;
  dueTodayCount: number;
}

export interface UserProfile {
  userId: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  priority?: ListTasksPriority;
  search?: string;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;