# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * To-Do List Application API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  highPriority: number;
  dueTodayCount: number;
}