# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * To-Do List Application API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
}