# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * To-Do List Application API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdatePriority } from "./taskUpdatePriority";
import type { TaskUpdateStatus } from "./taskUpdateStatus";

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
}