# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * To-Do List Application API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  userId: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
}